+++ src/FuelClient.cc
@@ -41,7 +41,7 @@ class ignition::fuel_tools::FuelClientPrivate
 {
   /// \brief A model URL.
   /// E.g.: https://api.ignitionfuel.org/1.0/caguero/models/Beer
-  public: const std::string kModelURLRegexStr{
+  public: const std::string kModelUrlRegexStr{
     // Method
     "^([[:alnum:]\\.\\+\\-]+):\\/\\/"
     // Server
@@ -110,7 +110,7 @@ FuelClient::FuelClient(const ClientConfig &_config, const REST &_rest,
     this->dataPtr->cache.reset(_cache);
 
   this->dataPtr->urlModelRegex.reset(new std::regex(
-    this->dataPtr->kModelURLRegexStr));
+    this->dataPtr->kModelUrlRegexStr));
   this->dataPtr->uniqueNameModelRegex.reset(new std::regex(
     this->dataPtr->kModelUniqueNameRegexStr));
 }
@@ -133,11 +133,11 @@ Result FuelClient::ModelDetails(const ServerConfig &/*_server*/,
   ignition::fuel_tools::REST rest;
   RESTResponse resp;
 
-  auto serverURL = _id.Server().URL();
+  auto serverUrl = _id.Server().URL();
   auto version = _id.Server().Version();
   auto path = ignition::common::joinPaths(_id.Owner(), "models", _id.Name());
 
-  resp = rest.Request(REST::GET, serverURL, version, path, {}, {}, "");
+  resp = rest.Request(REST::GET, serverUrl, version, path, {}, {}, "");
   if (resp.statusCode != 200)
     return Result(Result::FETCH_ERROR);
 
@@ -240,18 +240,27 @@ Result FuelClient::DeleteModel(const ServerConfig &/*_server*/,
 Result FuelClient::DownloadModel(const ServerConfig &/*_server*/,
   const ModelIdentifier &_id)
 {
+  // Server config
+  if (_id.Server().URL().empty() || _id.Server().Version().empty())
+  {
+    ignerr << "Can't download model, server configuration incomplete: "
+          << std::endl << _id.Server().AsString() << std::endl;
+    return Result(Result::FETCH_ERROR);
+  }
+
+  // Local path
+  auto path = ignition::common::joinPaths(_id.Owner(),
+      "models", _id.Name() + ".zip");
+
+  // Request
   ignition::fuel_tools::REST rest;
   RESTResponse resp;
-
-  auto serverURL = _id.Server().URL();
-  auto version = _id.Server().Version();
-  auto path = ignition::common::joinPaths(_id.Owner(), "models",
-    _id.Name() + ".zip");
-
-  resp = rest.Request(REST::GET, serverURL, version, path, {}, {}, "");
+  resp = rest.Request(REST::GET, _id.Server().URL(), _id.Server().Version(),
+      path, {}, {}, "");
   if (resp.statusCode != 200)
     return Result(Result::FETCH_ERROR);
 
+  // Save
   if (!this->dataPtr->cache->SaveModel(_id, resp.data, true))
     return Result(Result::FETCH_ERROR);
 
@@ -259,9 +268,21 @@ Result FuelClient::DownloadModel(const ServerConfig &/*_server*/,
 }
 
 //////////////////////////////////////////////////
-bool FuelClient::ParseModelURL(const std::string &_modelURL,
-  ServerConfig &/*_server*/, ModelIdentifier &_id)
+bool FuelClient::ParseModelURL(const std::string &_modelUrl,
+    ServerConfig &/*_server*/, ModelIdentifier &_id)
 {
+  return this->ParseModelUrl(common::URI(_modelUrl), _id);
+}
+
+//////////////////////////////////////////////////
+bool FuelClient::ParseModelUrl(const common::URI &_modelUrl,
+    ModelIdentifier &_id)
+{
+  if (!_modelUrl.Valid())
+    return false;
+
+  auto urlStr = _modelUrl.Str();
+
   std::smatch match;
   std::string scheme;
   std::string server;
@@ -270,7 +291,7 @@ bool FuelClient::ParseModelURL(const std::string &_modelURL,
   std::string name;
 
   // Try URL first
-  if (std::regex_match(_modelURL, match, *this->dataPtr->urlModelRegex) &&
+  if (std::regex_match(urlStr, match, *this->dataPtr->urlModelRegex) &&
       match.size() == 6u)
   {
     scheme = match[1];
@@ -280,7 +301,7 @@ bool FuelClient::ParseModelURL(const std::string &_modelURL,
     name = match[5];
   }
   // Then try unique name
-  else if (std::regex_match(_modelURL, match,
+  else if (std::regex_match(urlStr, match,
       *this->dataPtr->uniqueNameModelRegex) && match.size() == 5u)
   {
     scheme = match[1];
@@ -290,7 +311,7 @@ bool FuelClient::ParseModelURL(const std::string &_modelURL,
   }
   else
   {
-    ignerr << "Invalid URL [" << _modelURL << "]" << std::endl;
+    ignerr << "Invalid URL [" << urlStr << "]" << std::endl;
     return false;
   }
 
@@ -326,27 +347,55 @@ bool FuelClient::ParseModelURL(const std::string &_modelURL,
 }
 
 //////////////////////////////////////////////////
-Result FuelClient::DownloadModel(const std::string &_modelURL,
+Result FuelClient::DownloadModel(const std::string &_modelUrl,
   std::string &_path)
 {
+  return this->DownloadModel(common::URI(_modelUrl), _path);
+}
+
+//////////////////////////////////////////////////
+Result FuelClient::DownloadModel(const common::URI &_modelUrl,
+  std::string &_path)
+{
+  // Get data from URL
   ModelIdentifier id;
-  ServerConfig srv;
-  if (!this->ParseModelURL(_modelURL, srv, id))
+  if (!this->ParseModelUrl(_modelUrl, id))
   {
     return Result(Result::FETCH_ERROR);
   }
 
+  // Download
+  ServerConfig srv;
   auto result = this->DownloadModel(srv, id);
   if (result)
   {
-    // Convert name to lowercase.
-    std::string name = id.Name();
-    std::string owner = id.Owner();
-    std::transform(name.begin(), name.end(), name.begin(), ::tolower);
-    name = common::replaceAll(name, " ", "_");
     _path = ignition::common::joinPaths(this->Config().CacheLocation(),
-      "models", owner, name);
+        id.Server().Url().Path().Str(), id.Owner(), "models", id.Name());
   }
 
   return result;
 }
+
+//////////////////////////////////////////////////
+Result FuelClient::CachedModel(const common::URI &_modelUrl,
+  std::string &_path)
+{
+  // Get data from URL
+  ModelIdentifier id;
+  ServerConfig srv;
+  if (!this->ParseModelURL(_modelUrl.Str(), srv, id))
+  {
+    return Result(Result::FETCH_ERROR);
+  }
+
+  // Check local cache
+  auto modelIter = this->dataPtr->cache->MatchingModel(id);
+  if (modelIter)
+  {
+    _path = modelIter.PathToModel();
+    return Result(Result::FETCH_ALREADY_EXISTS);
+  }
+
+  return Result(Result::FETCH_ERROR);
+}
+